requirements.txt
django==4.1.2
requests
djangorestframework
django-jazzmin
Pillow